Add debug settings to avoid overwriting release installs
