+++ app-android/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
 
         <provider
             android:name="androidx.core.content.FileProvider"
-            android:authorities="com.lift.bro.fileprovider"
+            android:authorities="@string/file_provider_id"
             android:exported="false"
             android:grantUriPermissions="true">
             <meta-data
