+++ data/core/src/androidMain/kotlin/com/lift/bro/data/core/ai/NanoAIRepository.kt
@@ -10,7 +10,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.runBlocking
 
-class MLKitAIRepository : AIRepository {
+class NanoAIRepository : AIRepository {
 
     private val generativeModel by lazy { Generation.getClient() }
