+++ domain/src/commonMain/kotlin/com/lift/bro/domain/ai/DownloadAIModelUseCase.kt
@@ -1,0 +1,9 @@
+package com.lift.bro.domain.ai
+
+import com.lift.bro.domain.repositories.AIRepository
+
+class DownloadAIModelUseCase(
+    private val aiRepository: AIRepository
+) {
+    operator fun invoke() = aiRepository.downloadModel()
+}
