+++ domain/src/commonMain/kotlin/com/lift/bro/domain/ai/GetAIModelStatusUseCase.kt
@@ -1,0 +1,9 @@
+package com.lift.bro.domain.ai
+
+import com.lift.bro.domain.repositories.AIRepository
+
+class GetAIModelStatusUseCase(
+    private val aiRepository: AIRepository,
+) {
+    operator fun invoke() = aiRepository.getModelStatus()
+}
