+++ domain/src/commonMain/kotlin/com/lift/bro/domain/workout/WorkoutGenerationError.kt
@@ -1,0 +1,8 @@
+package com.lift.bro.domain.workout
+
+sealed class WorkoutGenerationError : Exception() {
+    data object ModelNotReady : WorkoutGenerationError()
+    data object GenerationFailed : WorkoutGenerationError()
+    data object ParseFailed : WorkoutGenerationError()
+}
+
