+++ presentation/compose/src/androidMain/kotlin/com/lift/bro/di/DependencyContainer.android.kt
@@ -12,7 +12,7 @@ import com.lift.bro.data.DriverFactory
 import com.lift.bro.data.LBDatabase
 import com.lift.bro.data.analytics.NoOpAnalytics
 import com.lift.bro.data.analytics.PostHogAnalytics
-import com.lift.bro.data.core.ai.MLKitAIRepository
+import com.lift.bro.data.core.ai.NanoAIRepository
 import com.lift.bro.data.datasource.UserPreferencesDataSource
 import com.lift.bro.data.repository.SettingsRepository
 import com.lift.bro.domain.analytics.Analytics
@@ -44,7 +44,7 @@ actual class DependencyContainer {
     }
 
     actual val aiRepository: AIRepository by lazy {
-        MLKitAIRepository()
+        NanoAIRepository()
     }
 
     actual fun launchUrl(url: String) {
